requirements.txt
discord.py
python-dotenv
fastapi
uvicorn
jinja2
httpx
itsdangerous
